ssl.c: move out X509 store APIs to new file
