+++ src/include.am
@@ -18,6 +18,7 @@ MAINTAINERCLEANFILES+= $(FIPS_FILES)
 
 EXTRA_DIST += src/bio.c
 EXTRA_DIST += src/conf.c
+EXTRA_DIST += src/x509_str.c
 
 $(FIPS_FILES):
 	$(AM_V_at)touch $(srcdir)/$@
